+++ Presentacion/Movimientos/formNuevoMovimiento.cs
@@ -334,7 +334,7 @@ namespace Presentacion
                             if (fila["codigo"].ToString().Equals(txtCodigo.Text))
                             {
                                 oCorteE.idCorte = Convert.ToInt32(fila["idCorte"].ToString());
-                                oCorteE.codigo = Convert.ToInt32(fila["codigo"].ToString());
+                                oCorteE.codigo = Convert.ToInt64(fila["codigo"].ToString());
                                 oCorteE.corte = fila["corte"].ToString();
                                 oCorteE.tipo = fila["tipo"].ToString();
                                 oCorteE.Promedio = Utilidades.Util_Form.convertFloat(fila["promedio"].ToString(), false);
